requirements.txt
# Core dependencies
alpaca-trade-api==3.1.1
pandas==2.1.4
numpy==1.26.2
PyYAML==6.0.1
requests==2.31.0
APScheduler==3.10.4
pytz==2023.3
Flask==3.0.0
pyngrok==7.0.1

# Testing dependencies (dev)
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
responses==0.24.1
